Exclusão da pasta orderTableRow, implementação de input type checkbox no componente card e função de trazer dados do pedido do firebase e printar na tela da cozinha
